+++ CHANGELOG.md
@@ -1,4 +1,13 @@
-# Staged
+# Mon Sep 17 20:48:18 UTC 2018
+
+* Remove Admob Video Caching AB Test
+* Add 广告 to Overlays for Tencent Ads
+* Move platform logic away from API class
+* Serializable campaign models
+* Parse new comet content types
+* Use a longer safeguard timeout for showCameraFeed
+
+# Mon Sep 17 08:00:27 UTC 2018
 
 * Enable new video overlay for all
 * Fix: VAST end screen not visible after video errors (black screen)